ui: Move get-component-factory into utils/dom
