+++ ui-v2/app/utils/dom/get-component-factory.js
